# src/main/java/ru/yegorsmirnov/testtask/controller/AuthController.java
package ru.yegorsmirnov.testtask.controller;


import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import ru.yegorsmirnov.testtask.model.dto.JwtAuthenticationResponse;
import ru.yegorsmirnov.testtask.model.dto.SignInRequest;
import ru.yegorsmirnov.testtask.model.dto.SignUpRequest;
import ru.yegorsmirnov.testtask.service.AuthenticationService;


@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
public class AuthController {
    private final AuthenticationService authenticationService;

    @PostMapping("/sign-up")
    public JwtAuthenticationResponse signUp(@RequestBody @Valid SignUpRequest request) {
        return authenticationService.signUp(request);
    }

    @PostMapping("/sign-in")
    public JwtAuthenticationResponse signIn(@RequestBody @Valid SignInRequest request) {
        return authenticationService.signIn(request);
    }
}